Refactor ImageCoursel to use TypeScript
These changes also migrate the component to be a functional component
and adds documentation to the props. There is a new scrollSpeed
prop that is optional but makes more semantic sense. That is used
to control scrolling speed but uses the old (now default) 7 second
scroll speed if none is specified.